Remove org.webcurator.ui classes.
Break the dependency of webcurator-core on the UI classes found in the
packages underneath org.webcurator.ui. This makes the project webcurator-core
more generic and less UI-centric.

We do this by removing the classes in org.webcurator.ui, as they belong in the
webcurator-webapp project. Move any classes and/or constants that are used
outside of the package org.webcurator.ui to org.webcurator.common and its
subpackages.

A corresponding patch import is done to webcurator-webapp to bring those ui
classes in with their history.